cambiar ints

+++ Tarea1_Pruebas.cpp
@@ -24,12 +24,12 @@ void impresionMetodos(vector<int> contOE){
 
     cout << count << endl;
 
-    /*
-    for (size_t i = 0; i < count; i++)
+    
+    for (int i = 0; i < count; i++)
     {
         cout << contOE[i] << endl;
     }
-    */
+
